migrate tests to pytest
